Create a standard project layout and move endpoints to the correct directories (#4)
* Create a standard project layout and move endpoints to the correct directories

* Separate structs into multiple files

* Move pointer for params and reponses to use objects

* Add Response struct to update and upgrade to server endpoint

* Use standard builder to build url and update CHANGELOG

* Move all structs to a type syntax

* Update function comments

* Update function comments
