+++ CHANGELOG.md
@@ -1,0 +1,22 @@
+# Changelog
+All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+
+## [v0.3.0] - 2022-12-07
+### Added
+- GitHub PR actions to run go vet, go lint and go mod tidy.
+
+### Updated
+- Updated project layout and structure.
+- Moved and refactored server and job api endpoints.
+- Updated our code to conform with Golang linter.
+- Updated our README to link to our Golang style and our license.
+- Upgraded golang v1.19.3.
+- Upgraded golangci-lint v1.50.1.
+
+## [v0.2.2] - 2022-05-20
+### Added
+- Added support for SSH keys when provisioning a server.
+
+## [v0.2.1] - 2022-05-16
+### Added
+- Added label update function.
