+++ pkg/api/job/doc.go
@@ -1,0 +1,2 @@
+// Package job represents our SiteHost `/job` endpoint
+package job
