+++ pkg/api/server/delete.go
@@ -1,0 +1,33 @@
+package server
+
+import (
+	"context"
+	"net/url"
+
+	"github.com/sitehostnz/gosh/pkg/utils"
+)
+
+// Delete a server with the provided name.
+func (s *Client) Delete(ctx context.Context, request DeleteRequest) (response DeleteResponse, err error) {
+	u := "server/delete.json"
+
+	keys := []string{
+		"client_id",
+		"name",
+	}
+
+	values := url.Values{}
+	values.Add("client_id", s.client.ClientID)
+	values.Add("name", request.Name)
+
+	req, err := s.client.NewRequest("POST", u, utils.Encode(values, keys))
+	if err != nil {
+		return response, err
+	}
+
+	if err := s.client.Do(ctx, req, &response); err != nil {
+		return response, err
+	}
+
+	return response, nil
+}
