+++ pkg/api/server/doc.go
@@ -1,0 +1,2 @@
+// Package server represents our SiteHost `/server` API endpoint.
+package server
