+++ pkg/models/doc.go
@@ -1,0 +1,2 @@
+// Package models hosts all sharing models between api endpoints.
+package models
