+++ pkg/utils/doc.go
@@ -1,0 +1,2 @@
+// Package utils contains our utilities functions.
+package utils
